requirements.txt
pywebview>=5.0
zstandard>=0.22.0
pyinstaller>=6.0.0
